refactor groundstations and satellites to share common code
